Allow to configure importance of Folders
This commit adds some logic that allows to manipulate the structure of
the generated tree by giving weights to certain folders. Elements with
higher weights will require less button presses to be reached.

As an example, for most of the year one might not be interested in
Christmas music. However, the "Favourites" folder always holds the most
favourite music and is accessed very often. Then, the custom weights
would be:

    {
     "Christmas": 0,
     "Favourites": 100
    }

Currently, all folders without configured weights get weight 1. Then,
the weights are renormalised to sum up to 1.

This commit also adds an exemplary JSON.