Delete setting of 'acpi=off'
In the boot configuration the ACPI capabilities were switched off. This
is a relict of weird stability problems encountered previously. It did
not help very much anyways.

Now, with a better SD card, it seems that this setting is not needed
anymore. Thus, it is removed to save some Watts of power and save some
components here and there.