Set access rights for cache directory
The whole script is designed such that it can be run as root while
MiniDLNA has its own user. Unfortunately, the "dlna_cache" directory had
not appropriate rights, which could lead to problems.

In this commit some lines are added that set the ownership and the
access rights of the directory at each invocation.