Stop to set owner/permissions due to hardlinks
No permissions are set on folders anymore. Since every file in DLNA/ is
either a hard or a soft link, setting permissions is tricky.

Some security issues require that the destination file of a hard link
must be read+write for the creator of that link. Therefore, we cannot
change the permission of music files in DLNA back to 660.

Furthermore, changing the owner recursively never should have worked,
again due to how hardlinks work.

All in all, all ownership/permission handling for DLNA/ had to be
removed. On second thought, dlna_cache does not need special care
neither; it is created and manipulated my minidlna exclusively.

It is assumed that these problems stem from a upgrade from
Debian/stretch to Ubuntu 18.04; the latter seems to be permissive.