Remove test that ensures balanced trees
For one, the test was not implemented correctly. It did not take into
account the length of the last split.

Furthermore, it is wrong to assume that a balanced tree is the optimal
solution. It is better to first fill individual subtrees, since then the
additional access costs can be avoided.